refactor(client): extract user tests to sibling file
